Move mocks from test files 🚚
